cbs_h2645: Improve handling of damaged content
Up until now, an error during decomposing of a fragment led to the
abortion of the decomposition and in the end, the whole fragment was
usually discarded.
Now an error during decomposition of an inessential unit (e.g. SEI units
for H2645) will mean that just the damaged unit will be discarded.