cbs: Add a macro to create functions for deep copying
Some structs in cbs (most importantly some of the parameter sets of
H.264/HEVC) can contain external buffers so that shallow copies are not
enough for them; furthermore, they need a special free-callback
function. The macro provided in this commit can be used to easily create
functions both for copying as well as freeing such structures.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@googlemail.com>