cbs: Add function to update video codec parameters
If any of the *_metadata filter based upon cbs currently updates a video
codec parameter like aspect ratio information or color information, the
AVCodecParameters are not updated accordingly, so that e.g. muxers write
wrong header values that may precede the new values on the bitstream
level. This commit adds a function to also update the video codec
parameters so that the above situation can be fixed in a unified manner.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@googlemail.com>