avcodec/cbs_bsf: Update AVCodecParameters
If any of the *_metadata filter based upon CBS currently updates a video
codec parameter like color information, the AVCodecParameters are not
updated accordingly, so that e.g. muxers write header values based upon
outdated information that may precede and thereby nullify the new values
on the bitstream level.

This commit therefore adds functionality to change this in a fairly
generic way: Fields corresponding to the properties that may potentially
be changed by one of the CBS BSFs have been added to CBSBSFContext.
They are initialized to invalid values and the BSF may modify them ad
libitum. Afterwards the fields set to valid values (set by the BSF) are
stored in the BSF's par_out.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>