av1_metadata: Add option to update video parameters
Up until now, this BSF only changed the bitstream, not the
AVCodecParameters. The result is that e.g. some muxers use outdated
information to write header information that conflicts with (and precedes)
the new information at the bitstream level, so that the updates might
not lead to the desired change.
This commit changes this. It adds a mechanism by which the new
information can be used to update the AVCodecParameters, too. This is
the new default and an option has been added to revert to the old
behaviour.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@googlemail.com>