hevc_metadata: Update AVCodecParameters
This commit adds an option to also update the AVCodecParameters to align
them with the changes at the bitstream level. It is on by default.

Given the importance of correct frame dimensions they are updated
even when the new option is disabled. Notice that for field-based HEVC
the derived height is the height of a frame, which deviates from current
FFmpeg behaviour.

This commit also adds documentation for the new value.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@googlemail.com>