startcode: Use common macro and switch to pointer arithmetic
The reasons are cosmetics and preparation for future patches that will
have even more cases and whose performance improves when switching to
direct pointer arithmetic: Benchmarks have shown that using pointers
directly instead of indexing to access the array to be about 5% faster
(33665 vs. 31806 for a 7.4 Mb/s H.264 file based on 10 iterations of
131072 runs each; and 244356 vs 233373 for a 30.2 Mb/s H.264 file based
on 10 iterations with 8192 runs each).

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>