av1_metadata: Update codec parameters
Up until now, this BSF only changed the bitstream and the extradata,
not the rest of the AVCodecParameters. The result is that e.g. some
muxers use outdated information to write header information that
conflicts with (and potentially precedes) the new information at the
bitstream level, so that using the bitstream filter might not have the
desired effect.

This commit changes this and thereby brings av1_metadata in line with
the expected behaviour for bitstream filters.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>