vp9_raw_reorder: Fix handling of frames that are never displayed
vp9_raw_reorder needs to cache several frames in order to know where to
insert a show_existing_frame-frame (if at all). It uses a conservative
approach to determine how many frames to cache: Frames are output if the
last of the reference frame slots occupied by a frame that still needs
to be displayed will be overwritten by a new frame or if a non-reference
frame (that needs to be displayed immediately) is encountered or during
flushing at EOF.

It is the first of these checks that this patch is about. It fails for
frames that will never be displayed, but are only intended as references
for other frames. So think of a situation like this: I0 ND P1 P2 ..., where
I is a keyframe with pts 0, ND is a never displayed frame and P1 and P2
are ordinary frames. No output is forced if (say) P2 overwrites the last
reference slot occupied by ND. In this case, ND will simply not be
output at all, corrupting the output.

The solution is to also check for whether the frame still needs to be
output.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>