vp9_raw_reorder: Fix memleaks and return value
Fixes the following issues:
1. Should an error happen during a write of a frame resulting from
encountering a non-reference frame, the non-reference frame leaked up
until now. Furthermore, the error was not forwarded in this case.
2. There can be situations in which ctx->next_frame is set after
vp9_raw_reorder_filter has returned (usually when the new frame resulted
in the output of an earlier frame). This frame hasn't been properly
freed when the bsf was closed or flushed.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>