vp9_raw_reorder: Output show_existing_frame if show_frame == 0
Whenever vp9_raw_reorder needs to decide which frame to output, it
determines two candidates for it: The earliest frame that has not yet
been output (ordered by coding order) and the earliest frame that still
needs to be displayed (ordered by pts). If these candidates coincide,
then it is presumed that no show_existing_frame is necessary for this
frame, i.e. the frame is treated as if show_frame for this frame were
set. And this is false, as there is no rule prohibiting to use a
combination of show_frame == 0 and an immediately following
show_existing_frame instead of a frame with show_frame == 1 in order to
output a frame immediately after it has been decoded.
It might even be necessary to do it that way if the frame in question
is a non-keyframe intra frame (show_frame frames can't be intra-only
non-keyframes by VP9's syntax). So check for show_frame == 1 instead of
for the coincidence of the two candidates.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>