mpeg2_metadata: Add feature to declare sequence progressive
It allows to switch the repeat_first_field and top_field_first flags
when the output is not declared a progressive sequence. It also allows
to declare coded frames to be progressive. And it allows to declare the
sequence a progressive_sequence if the right conditions are met.
Timing is left untouched.