matroskadec: Change handling of unknown-length levels
Unknown-length elements end when an element not allowed in them, but
allowed at a higher level is encountered. In order to check for this,
c1abd95a added a pointer to every syntax level's parent to each
EbmlSyntax. Given that the parent must of course also reference the
child in order to be able to enter said child level, one needs to use
forward declarations.
These forward declarations constitute tentative definitions and tentative
definitions with internal linkage (like our syntaxes) must not be an
incomplete type. Yet they were an incomplete type and while GCC and
Clang did not even warn about this (on default warning levels), it
broke compilation with MSVC.
Therefore this commit changes the system how unknown length levels are
handled: MatroskaLevel now includes a pointer to the parent syntax that
is set at runtime; therefore the syntax tables don't need to include a
pointer to their parents any more and so the forward declarations can be
removed.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>