h264_metadata: Import functionality of h264_redundant_pps
The version incorporated here is even more powerful: It allows the user
to specify the pic_init_qp_minus26 value of the output PPS. The LSB bit
of the qp-parameter tells whether weighted_pred_flag will be set to zero
or left untouched. The rest of qp consists of a six bit block that
contains a pic_init_qp_minus26 + 27 value.