avcodec/h264_metadata: Remove unused header parameter sets
h264_metadata now checks and informs the user if the extradata contains
parameter sets that are not referenced by any slice at all if the
extradata-mode is "minimize_2"; in this mode, new extradata that doesn't
contain the unneeded parameter sets any more is created at the end and
returned via a packet that only has an AV_PKT_DATA_NEW_EXTRADATA
side data entry. Whether muxers make use of this depends on the muxer.

"minimize_2" is also the new default mode.