h264_metadata: Add possibility to insert or minimize extradata
The "insert" modes add in-band extradata so that the track can be played
using in-band extradata alone from every keyframe on: "insert" adds all
currently visible extradata to every keyframe and also necessary
extradata to other frames if it changes mid-GOP. "insert_2" adds the
extradata belonging to the active SPS to every keyframe and also all
necessary extradata to other frames if it changes mid-GOP. "insert_3"
adds just the bare minimum: Only extradata that is actually used is
being inserted in-band.

"minimize" meanwhile tries to minimize the amount of in-band extradata
by using the extradata in the header as default: It is presumed that a
player will reset the extradata to the one contained in the header after
a seek.

Example: If a stream is divided into three parts A, B and C and A and
C use the extradata from the header, while B uses different (and
incompatible (i.e. the same SPS or PPS ids with different content))
extradata, then only the keyframes from B and the very first
keyframe from C need in-band extradata: Upon seeking to A or C, the
correct extradata is taken from the header (and in case of the first
keyframe from C immediately overwritten by the very same extradata).
Upon seeking to B, the incorrect extradata from the header is
immediately overwritten by the correct one, so that everything is fine.
During normal decoding, the correct extradata is always available
in-band at every point at which the needed extradata changes.

It is also possible to remove all in-band-extradata.