cbs_h2645: Implement copy-functions for parameter sets
and use them to replace the context's parameter sets.

This fixes a dangling pointers problem with the parameter sets
held in the CodedBitstreamH26xContext that occurs when the parameter
sets were not reference-counted initially.

The functions to free the parameter sets that have a special free
callback are now also created via the same macro as the copy functions.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@googlemail.com>