h264_metadata: Allow setting multiple qp values
This commit strengthens the ability to modify the qp values by allowing
the user to specify up to five qp values; for each of these values, a
PPS with this value will be created and the slices will be modified to
use the PPS so that the fewest bits are spent on the slice headers.

This mode is subject to the restriction that all PPS originally present
must have an id of 0 and after updating them, all these PPS must be
identical.

The LSB bit of the qp-parameter determines whether weighted_pred_flag
will be set to one or left untouched; the rest of qp consists of five
six bit blocks that contain a pic_init_qp_minus26 + 27 value, whereby
less serious blocks get a smaller pic_parameter_set_id. If such a block
is 0, it means not to insert such a PPS.