bytestream: Make get_bytes_left compatible with overread
bytestream2_get_bytes_left returns an unsigned int; as a result,
it returns big positive numbers if an overread already happened,
making it unsuitable for scenarios where one wants to allow this
(because the buffer is actually padded so that no segfaults can
happen). So add a function returning ptrdiff_t for this purpose;
given that it is intended to be used with the unsafe functions,
it has an "u" suffix, although it is not unsafe by itself.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>