h264_mp4toannexb: Improve extradata overread checks
Currently during parsing the extradata, h264_mp4toannexb checks for
overreads by adding the size of the current unit to the current position
pointer and comparing this to the end position of the extradata. But
pointer comparisons and pointer arithmetic is only defined if it does not
exceed the object it is used on (one past the last element of an array
is allowed, too). In practice, this might lead to overflows. Therefore
the check has been changed, using the new bytestream2_get_bytes_leftu
function.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>